requirements.txt
discord.py==2.3.2
yt-dlp
google-generativeai==0.8.3
python-dotenv==1.0.1
aiofiles==24.1.0
aiohttp==3.10.5